refactor: using InputEvent as a general key input design
`InputEvent` represents how a key input to the program now, it
indicates what key is from the input, and the status of other
functional keys includes ctrl, alt, shift and meta. It also includes
`raw` field which supports to write the representation into buffer.

Provide `CharKey` in enum to support `InputEvent` key fields to
indicate what key is actually the input event corresponds to.

Test cases are provided to show how the `InputEvent` is created from
bytes result.

`KeyCode` are kept for reference now.

 - Replace `KeyCode` into `InputEvent` in `parsing_byte` from input bytes
 - Extract common constants and log functions out